extracted styling out of SignInForm.js
